requirements.txt
fastapi
typesense
python-dotenv
uvicorn
